feat(#S11P21D109-243): nft API 작성 - 2

+++ backend/spring-server/src/main/java/com/semonemo/spring_server/domain/nft/dto/request/NFTServiceRequestDto.java
@@ -5,11 +5,13 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @NoArgsConstructor
 @ToString
 public class NFTServiceRequestDto {
     private Long userId;
-    private Long tokenId;
+    private BigInteger tokenId;
 }
